backend/requirement.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions

